fix: fix usage of PagerExperimental with 1 route and with useNativeDriver (#598)
### Motivation

If `PagerExperimental` is used on Android with 1 route and with `useNativeDriver` flag, then it doesn't render anything. Turned out that's because with 1 route `maxTranslate` calculates to `0`, and interpolation becomes
```
Animated.add(panX, offsetX).interpolate({
  inputRange: [-0, 0],
  outputRange: [-0, 0],
  extrapolate: 'clamp',
});
```
which doesn't work well with `useNativeDriver` (strangely enough, if one disables `useNativeDriver`, it works as expected).

Because in case with 1 route this calculates to non-changing animated `0` anyway, I've made this fix.

### Test plan

Use `PagerExperimental` on Android with 1 route and `useNativeDriver`. See that it renders content with my fix and doesn't render without it.