fix: memoize animated nodes to improve performance
Currently the animated nodes used for styles are re-created in render, which means that reanimated does a lot of expensive work detaching and reattaching them. Due to this, the framerate of the JS thread goes down to 0 FPS even on high-end devices when switching tabs quickly.

This PR implements memoization for the animated nodes so they are only recreated when necessary. On low-end devices, this keeps the JS frame rate around 12+ FPS on low-end devices and 40+ on high-end devices.

Related to #720