feat: add springVelocityScale effect (#839)
Motivation
In this PR I add springVelocityScale parameter as a config of Pager. Toss defines how much velocity of gesture impacts on an initial velocity of Pager. Found it very useful in bottom sheet.


However, while writing this code I figured out interesting bug in Pager which led to fact that initial velocity of spring was always zero. It happened bc previously velocity was set only if clock was not running. However few lines above we made similar check and if clock was not running... then we were starting it, so as the result there was no chance that in current state clock was not running.
I fixed it by moving starting clock few lines below.

Test plan
You may add springVelocityScale ={4} param to ScrollableTabBarExample